todoServer/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-encrypted-model-fields==0.6.5
django-simple-history==3.4.0
django-oauth-toolkit==2.3.0
social-auth-app-django==5.4.0
django-rest-framework-social-oauth2==1.1.0
channels==4.0.0
channels-redis==4.1.0
django-redis==5.4.0
pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
drf-nested-routers==0.93.4
coreapi==2.3.3
mysqlclient==2.2.6
python-dotenv==1.0.1
